homework08/requirements.txt
dash==4.1.0
dash_ag_grid
dash_bio
plotly.express
dash_bootstrap_components
Biopython
